Split nextest-filtering lib.rs content in sub-modules
